pyproject.toml
[tool.poetry]
name = "bluesky-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "bluesky_agent"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
atproto = "0.0.55" 
python-dotenv = "^1.0.0"
python-telegram-bot = "^21.9"
argparse = "^1.4.0"
black = "^24.10.0"
openai = "^1.58.1"
openbb = "^4.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"